refactor: DomainLayerのフォルダ構成の細分化
