+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/Repositories.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4ee098c5235ab4e4ebb2c073ff1abef7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
