+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/Repositories/ICommandHistory.cs.meta
