+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d163b6380e495b42841c2f8cae9d985
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
