+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/Services/ICommandParser.cs
@@ -1,7 +1,7 @@
 using System;
-using YukimaruGames.Terminal.Domain.Abstractions.Models;
+using YukimaruGames.Terminal.Domain.Abstractions.Models.ValueObjects;
 
-namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces
+namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Services
 {
     /// <summary>
     /// コマンド引数のパーサーインターフェイス.
