+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Interfaces/Services/ICommandRegistry.cs
@@ -1,6 +1,6 @@
-using YukimaruGames.Terminal.Domain.Abstractions.Models;
+using YukimaruGames.Terminal.Domain.Abstractions.Models.ValueObjects;
 
-namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces
+namespace YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Services
 {
     /// <summary>
     /// 登録されたコマンドの保存クラス.
