+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Models/Entities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cbdc02638eb5c5f4e9c7c6c0c9895709
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
