+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Models/ValueObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8eea4b46719bd642954f17b1b6ab185
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
