+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Abstractions/Models/ValueObjects/CommandArgument.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace YukimaruGames.Terminal.Domain.Abstractions.Models
+namespace YukimaruGames.Terminal.Domain.Abstractions.Models.ValueObjects
 {
     /// <summary>
     /// コマンドの引数型.
