+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandAutocomplete.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Services;
 
 namespace YukimaruGames.Terminal.Domain.Services
 {
