+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/CommandInvoker.cs
@@ -1,5 +1,5 @@
-using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
-using YukimaruGames.Terminal.Domain.Abstractions.Models;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Services;
+using YukimaruGames.Terminal.Domain.Abstractions.Models.ValueObjects;
 
 namespace YukimaruGames.Terminal.Domain.Services
 {
