+++ Assets/YukimaruGames/Terminal/Runtime/Domain/Services/Repositories.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ed270e1a1b1e4e418c3556aca379975
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
