+++ Assets/YukimaruGames/Terminal/Runtime/Runtime/Context/TerminalRuntimeScope.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using YukimaruGames.Terminal.Application.Interfaces;
-using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
+using YukimaruGames.Terminal.Domain.Abstractions.Interfaces.Services;
 
 namespace YukimaruGames.Terminal.Runtime
 {
